updated test component to city card modal
